moved writable functions into extensions module
