+++ grabbit/__init__.py
@@ -1,5 +1,6 @@
-from .core import (File, Entity, Layout, replace_entities, build_path,
-                   write_contents_to_file)
+from .core import File, Entity, Layout
+from .extensions import (replace_entities, build_path, write_contents_to_file,
+                         WritableFile, WritableLayout)
 
 __all__ = [
     'File',
@@ -7,5 +8,7 @@ __all__ = [
     'Layout',
     'replace_entities',
     'build_path',
-    'write_contents_to_file'
+    'write_contents_to_file',
+    'WritableFile',
+    'WritableLayout'
 ]
