+++ grabbit/extensions/__init__.py
@@ -1,5 +1,12 @@
-from hdfs import HDFSLayout
+# from .hdfs import HDFSLayout
+from .writable import (replace_entities, build_path, write_contents_to_file,
+                       WritableFile, WritableLayout)
 
 __all__ = [
-    'HDFSLayout'
+    # 'HDFSLayout',
+    'replace_entities',
+    'build_path',
+    'write_contents_to_file',
+    'WritableFile',
+    'WritableLayout'
 ]
